http://code.google.com/p/capirca/issues/detail?id=23
It seems reasonable to not render a 'protocol:: icmpv6' term inside an 'inet'
filter, and similarly not to render a 'protocol:: icmp' term inside an 'inet6'
filter.

You can now develop a filter that includes both icmp and icmp6 terms, which
only render when generating output of inet/inet6 respectively.  This should
improve the ability to make cross-platform / cross-AF policies.