Ignore warnings printed to stderr during smoke test.
The warning messages sent to logging appear to be cached, resulting in
global state being held during different test runs.  If the smoke test was run
independently, it passed, but when run as part of a group of tests,
it failed, as the warnings were getting generated and cached in the
earlier test, and so not generated/printed during the later run.

This isn't ideal, I would prefer to clear out the cached logged warning,
but it's not a big problem as these are just smoke tests anyway.