Generate output dir structure mirroring the input.
With this change, the input files

  - /input/policies/a.pol
  - /input/policies/subdir/b.pol

would yield the following files for output directory '/out':

  - /out/a.acl
  - /out/subdir/b.acl

Currently, all generated files are written to the same
directory if the output folder is an absolute path.  The output
for the example above is currently:

  - /out/a.acl
  - /out/b.acl

This is not good, as the 'subdir' structure in the input is lost,
and the same filename in a subdirectory could result in lost
output (e.g., suppose 'b.pol' above was 'a.pol' instead - both would
be written as '/out/a.acl', the last write would win).