Breaking change: specify includes relative to the policy base dir.
The user specifies the directory for the policy files. Prior to this change,
the 'include' specified 'policies' as the base for the includes, but that
results in the inclusion searching for '/parentdirs/policies/policies/'.

This change gives more consistency to the file structure.  I have assumed
that few people are taking advantage of the file inclusion feature, as it
is potentially confusing, and if they are using it, that it can be quickly
changed with sed/search-and-replace.