Add YamlNaming class and tests.
This change is similar to the YAML policy parser.  The definitions files
(networks and services) make sense to represent as YAML files, which standardizes
the format and somewhat simplifies parsing.

In ways the parsing for this is actually more complex as the YAML
commenting is slightly more user-friendly than the Capirca commenting.
For example, lists of items can be concatenated on to a single
line, and the comment placed at the end (in square brackets), eg:

  - a, b, c  [some comment]

There are sufficient tests such that this can be refactored in the future.