Added dummy symlinks to facilitate running all tests from a single command.
'python -m unittest discover' requires that all tests be named 'test_xxx.py'.
Renaming all of the 'tests' files to this naming convention would be extra thrash,
and perhaps would not be the right decision for the team/project.

This hack lets us run all of the tests in one shot using test discovery, at least
until the CI direction is set.