Add linter for policies and network/service defs
Summary: This adds linting support to Capirca. Linting is helpful to integrate to SDLC tooling and to ensure compliance to basic style guides, and can be used as an alternative to some of the common warnings when running aclgen. The README.md updates covers the use cases in more depth, as well as an example of the default style guide.

Test Plan: Ran linter against sample policies, boilerplate include file, and network/service definitions. Unit test offers some basic coverage on common linters.