Filecomment lexxer support
Summary:
It can be useful to sometimes embed arbitrary comments at the very top of
the file, rather than a few lines in via a header comment. This is supported
with a "file-comment" header attribute that on supported platforms will include
that content at the very top of the generated policy.

Test Plan:
Added new token to aclgenerator unit test - unittest did not
crash.