add __futures__ to all python files to support python2.7/3 code
Summary:

I'd like to add these future imports to all of the python code.
This will allow us to write python3 compatible code more easily, with the
eventualy hope that we can support python3 in the future.  The last big
hold out for not doing this was python-glfags not supporting python3. That
changed sometime last month so i'd like to get us moving in that direction.
The end is near.  https://pythonclock.org/

The harder part of this effort revolves around nacaddr.  We've solved this
once internally, but it could use some more work.  That effort will be in
later PR requests.

Test Plan: python2.7 -m unittest discover -p *_test.py

......................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
----------------------------------------------------------------------
Ran 518 tests in 82.943s

OK