Automated g4 rollback of changelist 244310350.
*** Reason for rollback ***

Rolling back, standard ipv6 ACLs are basically the same as extended. Looking into it more with regards to issues we found in other systems relying on the generators it seems that for all intents and purposes you use extended ACLs in this case. Inet6 does not append the keyword "extended" to the access-list creation as it does with inet. Unless a direct incompatibility between the way an "extended" inet6 is printed with how a "standard" inet6 ACL is accepted we have nothing to fix.

*** Original change description ***

Modifying cisco to allow ipv6 use in standard ACLs.
These are a bit different from IPv4 standard in that they require a non digit name.
https://www.cisco.com/c/en/us/td/docs/cable/cbr/configuration/guide/b_docsis_cbr_full_book_xe16_5/b_docsis_cbr_full_book_xe16_5_chapter_0111001.pdf

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=246332365