TypeError caused by sorting of v4 and v6 addresses in __str__ method of Policy class in policy.py.
Changes fix this sort so that it first sorts v4 and then v6, then concatenates the two sorted lists together.

The CL also removes the debug line. This is because outputting the whole policy object is truncated. The decision described in the bug, was to remove the output of this pol object.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=308101860