Add attribute calculation to GCE generator.
This allows users to pass in an integer that can be used to limit the max number of attributes in a VPC firewall policy. If a VPC firewall policy exceeds the max, an error will be thrown.

The library will now count the number of rules generated and log the number generated.

PiperOrigin-RevId: 347102880