Add check for ICMP types used with non-ICMP protocols.
Add unit tests for check above, and invalid ICMP type.

Background: this check was previously handled in
aclgenerator.Term.NormalizeIcmpTypes(), and the function call was
removed.  NormalizeIcmpTypes() is not address family "mixed" aware, so
just added the missing check needed.  And the local Term() class is
no longer needed, so remove it.