workflows: Upgrade actions to latest versions.
This updates actions to:

* `actions/checkout@v4`
* `actions/setup-python@v5`
* `actions/setup-node@v4`
* `actions/upload-artifact@v4`
* `microsoft/setup-msbuild@v1.3.3`

This is because of the following error which we have been receiving in the CI:

```
Error: This request has been automatically failed because it uses a deprecated version of `actions/upload-artifact: v2`. Learn more: https://github.blog/changelog/2024-02-13-deprecation-notice-v1-and-v2-of-the-artifact-actions/
```

This does not attempt to fix the pre-existing build issues caused by typing
problems introduced in other commits.